Moved paths to output.py among other functions to backend/archive
